Module Mcavl

An imperative, mutable, implementation of totally ordered Sets and Maps datastructures, with the following properties:

Thread-safe: concurrent updates from multiple domains are possible. These collections can typically be used to synchronize multiple workers towards a common goal.

Lock-free: concurrent modifications are able to make progress without waiting for others to finish.

Linearizable: concurrent operations appears to take place sequentially on a linearized timeline, providing a coherent collection of elements at all time. It is furthermore possible to take an O(1) snapshot/copy to observe its collection of elements at a given point in time.

module type Ordered = sig ... end

Totally ordered type.

module Set (Ord : Ordered) : sig ... end

Thread-safe mutable set structure given a totally ordered type.

module Map (Ord : Ordered) : sig ... end

Thread-safe mutable map structure given a totally ordered type.