Module Mcavl.Map

Thread-safe mutable map structure given a totally ordered type.

Parameters

module Ord : Ordered

Signature

type key = Ord.t

The type of the map keys.

type 'a t

The type of maps from type key to values of type 'a.

module View : sig ... end

A read-only, non-mutable view of a map (with more operations, compatible with the purely functional Stdlib.Map interface)

val empty : unit -> 'a t

empty () returns a new empty map. O(1)

val singleton : key -> 'a -> 'a t

singleton k v returns a new map containing only the binding k for v. O(1)

val add : key -> 'a -> 'a t -> unit

add k v t adds the binding k for v into the map t. If the key k was already bound, the previously bound value is replaced by v. O(logN)

val remove : key -> 'a t -> bool

remove k t deletes any binding of the key k in the map t. Returns true if the binding was removed, or false if no binding existed for this key. O(logN)

Queries

val is_empty : 'a t -> bool

is_empty t returns true when the map t contains no bindings, false if it has at least one. O(1)

val mem : key -> 'a t -> bool

mem k t returns true if the key k is bound in the map t, false otherwise. O(logN)

val min_binding : 'a t -> key * 'a

min_binding t returns the binding with the smallest key in the map t, or raises Not_found if the map is empty. O(logN)

val min_binding_opt : 'a t -> (key * 'a) option

min_binding t returns the binding with the smallest key in the map t, or None if the map is empty. O(logN)

val max_binding : 'a t -> key * 'a

max_binding t returns the binding with the largest key in the map t, or raises Not_found if the map is empty. O(logN)

val max_binding_opt : 'a t -> (key * 'a) option

max_binding_opt t returns the binding with the largest key in the map t, or None if the map is empty. O(logN)

val choose : 'a t -> key * 'a

choose t returns an arbitrary binding of the map t, or raises Not_found if the map is empty. O(1)

val choose_opt : 'a t -> (key * 'a) option

choose_opt t returns an arbitrary binding of the map t, or None if the map is empty. O(1)

val find : key -> 'a t -> 'a

find k t returns the value associated with the key k in the map t, or raises Not_found if no such binding existed. O(logN)

val find_opt : key -> 'a t -> 'a option

find_opt k t returns the value associated with the key k in the map t, or returns None if no such binding existed. O(logN)

val find_first : ( key -> bool ) -> 'a t -> key * 'a

find_first predicate t returns the smallest binding of the map t that satisfies that monotonically increasing predicate, or raises Not_found if no such binding exists. O(logN)

val find_first_opt : ( key -> bool ) -> 'a t -> (key * 'a) option

find_first_opt predicate t returns the smallest binding of the map t that satisfies that monotonically increasing predicate, or None if no such binding exists. O(logN)

val find_last : ( key -> bool ) -> 'a t -> key * 'a

find_last predicate t returns the largest binding of the map t that satisfies that monotonically increasing predicate, or raises Not_found if no such binding exists. O(logN)

val find_last_opt : ( key -> bool ) -> 'a t -> (key * 'a) option

find_last_opt predicate t returns the largest binding of the map t that satisfies that monotonically increasing predicate, or None if no such binding exists. O(logN)

Snapshots

Concurrent modifications of a map are linearizable. The snapshot/copy provides a coherent view of the bindings of a map along this linearized timeline.

Further updates to the original map (or its copies) will trigger a minimal copy-on-write of the internal substructures of the map. This doesn't impact the time complexity of any operations, but induces a corresponding memory complexity for copying the modified subparts once.

val copy : 'a t -> 'a t

copy t returns an independently mutable copy of the map t. Further modifications of the map t will not affect its copies (and vice-versa.) O(1)

val snapshot : 'a t -> 'a View.t

snapshot t returns a read-only view of the bindings of the map t. O(1)

val to_view : 'a t -> 'a View.t

Same as snapshot.

val of_view : 'a View.t -> 'a t

of_view v returns a new mutable map containing all the bindings of the view v. O(1)

Iterators

The following functions all proceed on a coherent snapshot of their map t argument, created at the start of their execution. Concurrent modifications of the original map t during their execution will not be observed by these traversals.

val cardinal : 'a t -> int

cardinal t returns the number of bindings in the map t. O(N)

val fold : ( key -> 'a -> 'b -> 'b ) -> 'a t -> 'b -> 'b

fold f t init computes f k0 v0 init |> f k1 v1 |> ... |> f kN vN, where k0,v0 ... kN,vN are the bindings of the map t in increasing order. Returns init if the map t was empty.

val iter : ( key -> 'a -> unit ) -> 'a t -> unit

iter f t calls f on all the bindings of the set t in increasing order.

val for_all : ( key -> 'a -> bool ) -> 'a t -> bool

for_all predicate t returns true when all the bindings of the map t satisfies the predicate. No ordering is guaranteed and the function will exit early if it finds an invalid binding.

val exists : ( key -> 'a -> bool ) -> 'a t -> bool

exists predicate t returns true when at least one binding of the map t satisfies the predicate. No ordering is guaranteed and the function will exit early if it finds a valid binding.

val bindings : 'a t -> (key * 'a) list

bindings t is the list of all the bindings of t in increasing order.

val to_list : 'a t -> (key * 'a) list

Same as bindings.

val of_list : (key * 'a) list -> 'a t

of_list lst is the map containing all the bindings of the list lst.

val to_seq : 'a t -> (key * 'a) Stdlib.Seq.t

to_seq t is the sequence containing all the bindings of the map t in increasing order. O(1) creation then O(1) amortized for each consumed binding of the sequence (with O(logN) worst-case)

val to_rev_seq : 'a t -> (key * 'a) Stdlib.Seq.t

to_rev_seq t is the sequence containing all the bindings of the map t in decreasing order.

val to_seq_from : key -> 'a t -> (key * 'a) Stdlib.Seq.t

to_seq_from k t is the sequence containing all the bindings of the set t whose key is larger than or equal to k, in increasing order.

val of_seq : (key * 'a) Stdlib.Seq.t -> 'a t

of_set seq is the map containing all the bindings of the sequence seq.