Module Map.View

A read-only, non-mutable view of a map (with more operations, compatible with the purely functional Stdlib.Map interface)

type key = Ord.t

The type of the map keys.

type 'a t

The type of maps from type key to values of type 'a.

val empty : 'a t

The empty map.

val singleton : key -> 'a -> 'a t

singleton k v returns a map containing only the binding k for v. O(1)

val add : key -> 'a -> 'a t -> 'a t

add k v t returns a map containing the binding k for v and all the bindings of t. If k was already bound in the map t, then its previous value is replaced by v. O(logN)

val remove : key -> 'a t -> 'a t

remove k t returns a map containing the bindings of t without k. If k was not bound in the map t, then the result is physically equal to t. O(logN)

val union : 'a t -> 'a t -> 'a t

union t1 t2 returns a map containing all the bindings of t1 and t2. O(N) worst-case

val inter : 'a t -> 'a t -> 'a t

inter t1 t2 returns a map containing the shared bindings of t1 and t2. O(N) worst-case

val diff : 'a t -> 'a t -> 'a t

diff t1 t2 returns a map containing the bindings of t1 whose keys are not bound in the map t2. O(N) worst-case

val map : ( 'a -> 'a ) -> 'a t -> 'a t

map f t returns a map where each bound value in the map t has been replaced by its mapping by f:

The values are passed to f in increasing key order.

The result is physically equal to t if f always returned a physically equal element. O(NlogN) worst-case

val mapi : ( key -> 'a -> 'a ) -> 'a t -> 'a t

mapi f t returns a map where each bound value in the map t has been replaced by its mapping by f:

The values are passed to f in increasing key order.

The result is physically equal to t if f always returned a physically equal element. O(NlogN) worst-case

val filter : ( key -> 'a -> bool ) -> 'a t -> 'a t

filter predicate t returns the subset of bindings of the map t that satistifies the predicate (called in increasing key order). The resulting map is physical equal to t if no binding was rejected. O(N)

val filter_map : ( key -> 'a -> 'a option ) -> 'a t -> 'a t

filter_map predicate t returns a map containing the Some bindings of f x0, f x1, ..., f xN where x0 ... xN are all the elements of the map t.

The bindings are passed to f in increasing order.

The result is physically equal to t if f always returned Some physically equal value. O(NlogN) worst-case

val partition : ( key -> 'a -> bool ) -> 'a t -> 'a t * 'a t

partiton predicate t returns two maps, the first one containing all the bindings of the map t that satisfies predicate, while the second contains all the rejected ones.

The bindings are passed to f in increasing order.

The first map is physically equal to t if f always returned true

The second map is physically equal to t if f always returned false. O(N)

val split : key -> 'a t -> 'a t * 'a option * 'a t

split k t returns a triple (smaller, found, larger) such that:

smaller is the subset of bindings whose key is strictly smaller than k

larger is the subset of bindings whose key is strictly larger than k

found is Some v if k is bound to v in the map t, None otherwise. O(logN)

val pop_min : 'a t -> key * 'a * 'a t

pop_min t returns the smallest binding and the other bindings of the map t, or raises Not_found if the map t is empty. O(logN)

val pop_min_opt : 'a t -> (key * 'a * 'a t) option

pop_min_opt t returns the smallest binding and the other bindings of the map t, or None if the map t is empty. O(logN)

val pop_max : 'a t -> key * 'a * 'a t

pop_max t returns the largest binding and the other bindings of the map t, or raises Not_found if the map t is empty. O(logN)

val pop_max_opt : 'a t -> (key * 'a * 'a t) option

pop_max_opt t returns the largest binding and the other bindings of the map t, or None if the map t is empty.

Queries

val is_empty : 'a t -> bool

is_empty t returns true when the map t contains no bindings, false if it has at least one. O(1)

val mem : key -> 'a t -> bool

mem k t returns true if the key k is bound in the map t, false otherwise. O(logN)

val min_binding : 'a t -> key * 'a

min_binding t returns the binding with the smallest key in the map t, or raises Not_found if the map is empty. O(logN)

val min_binding_opt : 'a t -> (key * 'a) option

min_binding t returns the binding with the smallest key in the map t, or None if the map is empty. O(logN)

val max_binding : 'a t -> key * 'a

max_binding t returns the binding with the largest key in the map t, or raises Not_found if the map is empty. O(logN)

val max_binding_opt : 'a t -> (key * 'a) option

max_binding_opt t returns the binding with the largest key in the map t, or None if the map is empty. O(logN)

val choose : 'a t -> key * 'a

choose t returns an arbitrary binding of the map t, or raises Not_found if the map is empty. O(1)

val choose_opt : 'a t -> (key * 'a) option

choose_opt t returns an arbitrary binding of the map t, or None if the map is empty. O(1)

val find : key -> 'a t -> 'a

find k t returns the value associated with the key k in the map t, or raises Not_found if no such binding existed. O(logN)

val find_opt : key -> 'a t -> 'a option

find_opt k t returns the value associated with the key k in the map t, or returns None if no such binding existed. O(logN)

val find_first : ( key -> bool ) -> 'a t -> key * 'a

find_first predicate t returns the smallest binding of the map t that satisfies that monotonically increasing predicate, or raises Not_found if no such binding exists. O(logN)

val find_first_opt : ( key -> bool ) -> 'a t -> (key * 'a) option

find_first_opt predicate t returns the smallest binding of the map t that satisfies that monotonically increasing predicate, or None if no such binding exists. O(logN)

val find_last : ( key -> bool ) -> 'a t -> key * 'a

find_last predicate t returns the largest binding of the map t that satisfies that monotonically increasing predicate, or raises Not_found if no such binding exists. O(logN)

val find_last_opt : ( key -> bool ) -> 'a t -> (key * 'a) option

find_last_opt predicate t returns the largest binding of the map t that satisfies that monotonically increasing predicate, or None if no such binding exists. O(logN)

Iterators

val cardinal : 'a t -> int

cardinal t returns the number of bindings in the map t. O(N)

val fold : ( key -> 'a -> 'b -> 'b ) -> 'a t -> 'b -> 'b

fold f t init computes f k0 v0 init |> f k1 v1 |> ... |> f kN vN, where k0,v0 ... kN,vN are the bindings of the map t in increasing order. Returns init if the map t was empty.

val iter : ( key -> 'a -> unit ) -> 'a t -> unit

iter f t calls f on all the bindings of the set t in increasing order.

val for_all : ( key -> 'a -> bool ) -> 'a t -> bool

for_all predicate t returns true when all the bindings of the map t satisfies the predicate. No ordering is guaranteed and the function will exit early if it finds an invalid binding.

val exists : ( key -> 'a -> bool ) -> 'a t -> bool

exists predicate t returns true when at least one binding of the map t satisfies the predicate. No ordering is guaranteed and the function will exit early if it finds a valid binding.

val bindings : 'a t -> (key * 'a) list

bindings t is the list of all the bindings of t in increasing order.

val to_list : 'a t -> (key * 'a) list

Same as bindings.

val of_list : (key * 'a) list -> 'a t

of_list lst is the map containing all the bindings of the list lst.

val to_seq : 'a t -> (key * 'a) Stdlib.Seq.t

to_seq t is the sequence containing all the bindings of the map t in increasing order. O(1) creation then O(1) amortized for each consumed binding of the sequence (with O(logN) worst-case)

val to_rev_seq : 'a t -> (key * 'a) Stdlib.Seq.t

to_rev_seq t is the sequence containing all the bindings of the map t in decreasing order.

val to_seq_from : key -> 'a t -> (key * 'a) Stdlib.Seq.t

to_seq_from k t is the sequence containing all the bindings of the set t whose key is larger than or equal to k, in increasing order.

val of_seq : (key * 'a) Stdlib.Seq.t -> 'a t

of_set seq is the map containing all the bindings of the sequence seq.