Module type Mcavl.Ordered

Totally ordered type.

type t

The type of comparable elements.

val compare : t -> t -> int

compare a b must return 0 if a equals b, a negative number if a is stricly less than b and a positive number if a is strictly larger than b.